# Lab2/ark-pzpi-23-3-osychenko-illia-lab2/docs/swagger.yaml
basePath: /api
definitions:
  handler.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      role_id:
        type: integer
    required:
    - email
    - full_name
    - password
    - role_id
    type: object
  handler.DeviceAuthRequest:
    properties:
      serial_number:
        type: string
      token:
        type: string
    required:
    - serial_number
    - token
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handler.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  handler.UpdateUserRoleRequest:
    properties:
      role_id:
        type: integer
    required:
    - role_id
    type: object
...
  model.Permission:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
    type: object
...
  model.User:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      full_name:
        example: Іван Іванов
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      role:
        $ref: '#/definitions/model.Role'
      role_id:
        example: 2
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  service.DeviceAuthResponse:
    properties:
      access_token:
        type: string
      device_id:
        type: integer
      expires_in:
        type: integer
    type: object
  service.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@busoptima.ua
    name: API Support
  description: API для системи оптимізації автобусних перевезень BusOptima
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BusOptima API
  version: "1.0"
paths:
  /admin/audit-logs:
    get:
      consumes:
      - application/json
      description: Повертає записи журналу аудиту з можливістю пагінації
      parameters:
      - default: 1
        description: Номер сторінки
        in: query
        name: page
        type: integer
      - default: 20
        description: Кількість записів на сторінці
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати журнал аудиту
      tags:
      - Admin
  /admin/backup:
    post:
      consumes:
      - application/json
      description: Створює резервну копію бази даних
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Створити резервну копію
      tags:
      - Admin
  /admin/users:
    get:
      consumes:
      - application/json
      description: Повертає список всіх користувачів системи
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати список користувачів
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Створює нового користувача в системі
      parameters:
      - description: Дані користувача
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Створити нового користувача
      tags:
      - Admin
  /admin/users/{id}:
    put:
      consumes:
      - application/json
      description: Оновлює існуючого користувача за ID
      parameters:
      - description: ID користувача
        in: path
        name: id
        required: true
        type: integer
      - description: Оновлені дані користувача
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оновити користувача
      tags:
      - Admin
  /admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Оновлює роль користувача за ID
      parameters:
      - description: ID користувача
        in: path
        name: id
        required: true
        type: integer
      - description: Нова роль користувача
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оновити роль користувача
      tags:
      - Admin
  /analytics/dashboard:
    get:
      consumes:
      - application/json
      description: Повертає основні метрики та статистику для дашборду
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати дані дашборду
      tags:
      - Analytics
  /analytics/forecast:
    get:
      consumes:
      - application/json
      description: Повертає прогноз попиту на маршрути з аналітикою
      parameters:
      - description: ID маршруту для прогнозу
        in: query
        name: route_id
        type: integer
      - default: 7
        description: Кількість днів для прогнозу
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати прогноз попиту
      tags:
      - Analytics
  /analytics/profitability:
    get:
      consumes:
      - application/json
      description: Повертає детальний аналіз прибутковості маршрутів та рейсів
      parameters:
      - description: Дата початку періоду (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: Дата кінця періоду (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: ID маршруту для фільтрації
        in: query
        name: route_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Отримати аналіз прибутковості
      tags:
      - Analytics
  /auth/device:
    post:
      consumes:
      - application/json
      description: Автентифікує IoT-пристрій та повертає токен доступу
      parameters:
      - description: Дані пристрою
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/handler.DeviceAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.DeviceAuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Автентифікація IoT-пристрою
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Автентифікує користувача та повертає JWT токени
      parameters:
      - description: Облікові дані користувача
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Автентифікація користувача
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Оновлює токен доступу за допомогою refresh токена
      parameters:
      - description: Refresh токен
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Оновити токен доступу
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"